Move test_no_connection_state from test_cover to test_init
